# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ComboFinder API — phone display combo compatibility
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Stats {
  totalBrands: number;
  totalModels: number;
  totalCombos: number;
}

export interface Brand {
  id: number;
  name: string;
  logoUrl?: string | null;
  modelCount: number;
  createdAt: string;
}

export interface CreateBrandInput {
  name: string;
  logoUrl?: string | null;
}

export interface Model {
  id: number;
  brandId: number;
  brandName: string;
  name: string;
  releaseYear?: number | null;
  imageUrl?: string | null;
  comboCount: number;
  createdAt: string;
}

export type ComboComboType = typeof ComboComboType[keyof typeof ComboComboType];


export const ComboComboType = {
  OEM: 'OEM',
  Compatible: 'Compatible',
  Refurbished: 'Refurbished',
} as const;

export interface Combo {
  id: number;
  modelId: number;
  modelName: string;
  brandName: string;
  name: string;
  comboType: ComboComboType;
  qualityGrade?: string | null;
  notes?: string | null;
  priceRange?: string | null;
  inStock: boolean;
  createdAt: string;
}

export interface ModelWithCombos {
  id: number;
  brandId: number;
  brandName: string;
  name: string;
  releaseYear?: number | null;
  imageUrl?: string | null;
  combos: Combo[];
  createdAt: string;
}

export interface CreateModelInput {
  brandId: number;
  name: string;
  releaseYear?: number | null;
  imageUrl?: string | null;
}

export type CreateComboInputComboType = typeof CreateComboInputComboType[keyof typeof CreateComboInputComboType];


export const CreateComboInputComboType = {
  OEM: 'OEM',
  Compatible: 'Compatible',
  Refurbished: 'Refurbished',
} as const;

export interface CreateComboInput {
  modelId: number;
  name: string;
  comboType: CreateComboInputComboType;
  qualityGrade?: string | null;
  notes?: string | null;
  priceRange?: string | null;
  inStock: boolean;
}

export interface SearchResults {
  brands: Brand[];
  models: Model[];
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type SearchModelsParams = {
q?: string;
brand_id?: number;
};

export type GetModelsParams = {
brand_id?: number;
};

export type GetCombosParams = {
model_id?: number;
};