# lib/api-zod/src/generated/types/combo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ComboFinder API — phone display combo compatibility
 * OpenAPI spec version: 0.1.0
 */
import type { ComboComboType } from './comboComboType';

export interface Combo {
  id: number;
  modelId: number;
  modelName: string;
  brandName: string;
  name: string;
  comboType: ComboComboType;
  qualityGrade?: string | null;
  notes?: string | null;
  priceRange?: string | null;
  inStock: boolean;
  createdAt: Date;
}