# lib/api-zod/src/generated/types/createBrandInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ComboFinder API — phone display combo compatibility
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBrandInput {
  name: string;
  logoUrl?: string | null;
}