# lib/api-zod/src/generated/types/createComboInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ComboFinder API — phone display combo compatibility
 * OpenAPI spec version: 0.1.0
 */
import type { CreateComboInputComboType } from './createComboInputComboType';

export interface CreateComboInput {
  modelId: number;
  name: string;
  comboType: CreateComboInputComboType;
  qualityGrade?: string | null;
  notes?: string | null;
  priceRange?: string | null;
  inStock: boolean;
}