# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ComboFinder API — phone display combo compatibility
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  error: string;
}