# lib/api-zod/src/generated/types/searchResults.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ComboFinder API — phone display combo compatibility
 * OpenAPI spec version: 0.1.0
 */
import type { Brand } from './brand';
import type { Model } from './model';

export interface SearchResults {
  brands: Brand[];
  models: Model[];
}